+++ CHANGELOG.md
@@ -3,6 +3,7 @@ next release / 2014-02-??
 * Dropped sha256 and ripemd160 deps. Upgraded to crypto-hashing library.
 * removed function wrapper
 * Changed the way the constuctor works. Only supports input types of `Array`, `Buffer`, or `Uint8Array`. Does NOT randomly generate a private key anymore.
+* added `publicKey` property
 
 0.3.0 / 2014-02-03
 ------------------